refactor: se actualizan librerías y limpieza de código
