+++ components/AvatarChip.jsx
