+++ components/UsedInCraft.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import Typography from "@mui/material/Typography";
-import { Link } from "@mui/material";
-import { errorHandler } from "../reducers/apiReducer.js";
-import { useNotificationContext } from "../context/notification";
+import Link from "next/link";
+import { errorHandler } from "@reducers/apiReducer.js";
+import { useNotificationContext } from "@context/notification.jsx";
 
 export default function UsedInCraft({ craftsUsing }) {
   // eslint-disable-next-line  no-unused-vars
@@ -17,13 +17,7 @@ export default function UsedInCraft({ craftsUsing }) {
             color="text.primary"
             gutterBottom
           >
-            <Link
-              onClick={() => {
-                errorHandler(dispatcher, "Error desde UsedInCraft");
-              }}
-            >
-              {item.name} - {station.name} [{level}]
-            </Link>
+            {item.name} - {station.name} [{level}]
           </Typography>
         ))
       )}
