+++ package.json
@@ -12,10 +12,11 @@
     "@mui/lab": "^5.0.0-alpha.116",
     "@mui/material": "^5.11.5",
     "axios": "^1.2.3",
+    "eslint-config-next": "^14.2.5",
     "graphql-request": "^5.1.0",
-    "next": "latest",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "next": "^14.2.5",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "swr": "^2.0.1"
   },
   "devDependencies": {
